Revert "Dev bug fix"
