+++ src/ChinookTest.cpp
@@ -4,6 +4,52 @@
 /* settings */
 const static double global_speed_limit = 0.8;
 
+
+/* utils */
+void reportMsg(std::string msg){
+    FRC_NetworkCommunication_sendConsoleLine(msg.c_str());
+}
+
+enum {
+    // wheels
+    LeftPortFront = 0,
+    LeftPortBack = 2,
+    RightPortFront = 1,
+    RightPortBack = 3,
+    // Ball pick up
+    LeftBall = 4,
+    RightBall = 5,
+    // Gamepad Port Mappings
+    GamepadPort = 3,
+    GamepadAButton = 1,
+    GamepadBButton = 2,
+    GamepadXButton = 3,
+    GamepadYButton = 4,
+    GamepadLeftBumper = 5,
+    GamepadRightBumper = 6,
+    GamepadBack = 7,
+    GamepadStart = 8,
+    GamepadLeftStickDown = 1,
+    GamepadRightStickDown = 10,
+    JoystickTrigger = 1,
+    // Max Power
+    MaxPower = 2,
+    // Axis
+    LeftStickX = 0,
+    LeftStickY = 1,
+    LeftTrigger = 2,
+    RightTrigger = 3,
+    RightStickX = 4,
+    RightStickY = 5,
+    // Actuators
+    BorisPort = 8,
+    // Joystick
+    JoystickPort = 3,
+    JoystickAButton = 3,
+    JoystickYButton = 4,
+    JoystickLeftStickDown = 1
+};
+
 struct {
     // button
     bool jsIsAButtonPressed = false;
@@ -103,6 +149,64 @@ void resetOpState() {
     // DPAD
     OpState.gpDpad = -1.0;
 }
+
+class DriveTrain {
+public:
+    /**
+     * Constructor for DriveTrain
+     * @param speedLimit double The Speed Limit
+     */
+    explicit DriveTrain(double speedLimit) : speedLimit(speedLimit),
+                                    leftFront(LeftPortFront),
+                                    leftBack(LeftPortBack),
+                                    rightFront(RightPortFront),
+                                    rightBack(RightPortBack),
+                                    drive(leftFront, leftBack, rightFront, rightBack) {}
+
+    /**
+     * ArcadeDrive method
+     * @param x double rotation
+     * @param y double drive
+     */
+    void ArcadeDrive(double x, double y) { drive.ArcadeDrive(-limitSpeed(y), -(limitSpeed(std::abs(y) >= 0.5 ? x * 0.6 : x))); }
+
+    /**
+     * TankDrive method
+     * @param left double left-stick
+     * @param right double right-stick
+     */
+    void TankDrive(double left, double right) { drive.TankDrive(limitSpeed(left), limitSpeed(right)); }
+
+    /**
+     * Updates the speed limit
+     * @param speedLimit double renews speed limit to a new one
+     */
+    void Update(double speedLimit) { this->speedLimit = speedLimit; }
+
+    /**
+     * Change by a speed limit
+     * @param speedLimit  double increases/decreases speedLimit
+     */
+    void ChangeBy(double speedLimit) {
+        this->speedLimit += ((this->speedLimit + speedLimit > 1.0000000001) ||
+                (this->speedLimit + speedLimit < -0.0000000001) ) ? 0: speedLimit;
+        reportMsg("\rNewSpeedLimit" + std::to_string(this->speedLimit) + "     ");
+        std::cout << "New Limit: " << this->speedLimit << std::endl;
+    }
+
+private:
+    double speedLimit;
+    Victor leftFront, leftBack, rightFront, rightBack;
+    RobotDrive drive;
+
+    /**
+     * The devised algorithm for limiting the speed.
+     * @param a
+     * @return
+     */
+    double limitSpeed(double a) { return a/std::abs(a) * (std::abs(a) > speedLimit ? speedLimit : std::abs(a)); }
+
+};
 
 class Robot: public frc::IterativeRobot {
 public:
