--- src/Constant.h
@@ -1,44 +1,0 @@
-#ifndef CONSTANT_H
-#define CONSTANT_H
-
-enum {
-    // wheels
-    LeftPortFront = 0,
-    LeftPortBack = 2,
-    RightPortFront = 1,
-    RightPortBack = 3,
-    // Ball pick up
-    LeftBall = 4,
-    RightBall = 5,
-    // Gamepad Port Mappings
-    GamepadPort = 3,
-    GamepadAButton = 1,
-    GamepadBButton = 2,
-    GamepadXButton = 3,
-    GamepadYButton = 4,
-    GamepadLeftBumper = 5,
-    GamepadRightBumper = 6,
-    GamepadBack = 7,
-    GamepadStart = 8,
-    GamepadLeftStickDown = 1,
-    GamepadRightStickDown = 10,
-    JoystickTrigger = 1,
-    // Max Power
-    MaxPower = 2,
-    // Axis
-    LeftStickX = 0,
-    LeftStickY = 1,
-    LeftTrigger = 2,
-    RightTrigger = 3,
-    RightStickX = 4,
-    RightStickY = 5,
-    // Actuators
-    BorisPort = 8,
-    // Joystick
-    JoystickPort = 3,
-    JoystickAButton = 3,
-    JoystickYButton = 4,
-    JoystickLeftStickDown = 1
-};
-
-#endif /* Constant.h */
