--- src/util.cpp
@@ -1,4 +1,0 @@
-#include "util.h"
-
-/* utils */
-
