--- src/util.h
@@ -1,7 +1,0 @@
-#ifndef UTIL_H
-#define UTIL_H
-
-#include "ChinookTest.h"
-//inline void reportMsg(std::string msg);
-
-#endif /* Util.h */
